feat: add beneficiary, station and station-employe

+++ backend/src/company/company.service.ts
@@ -1,4 +1,8 @@
-import { ConflictException, Injectable } from '@nestjs/common';
+import {
+  ConflictException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { CreateCompanyDto } from './dto/create-company.dto';
 import { UpdateCompanyDto } from './dto/update-company.dto';
 import { PrismaService } from 'src/prisma/prisma.service';
@@ -9,8 +13,7 @@ export class CompanyService {
 
   async create(data: CreateCompanyDto) {
     const company = await this.findByName(data.name);
-    if (company)
-      throw new ConflictException(`L'entreprise ${data.name} existe déjà`);
+    if (company) throw new ConflictException(`${data.name} already exists!`);
 
     return await this.prisma.company.create({ data });
   }
@@ -20,18 +23,30 @@ export class CompanyService {
   }
 
   async findById(id: string) {
-    return await this.prisma.company.findUnique({ where: { id } });
+    const company = await this.prisma.company.findUnique({ where: { id } });
+
+    if (!company) throw new NotFoundException('Company not found!');
+
+    return company;
   }
 
   async findByName(name: string) {
-    return await this.prisma.company.findUnique({ where: { name } });
+    const company = await this.prisma.company.findUnique({ where: { name } });
+
+    if (!company) throw new NotFoundException('Company not found!');
+
+    return company;
   }
 
   async update(id: string, data: UpdateCompanyDto) {
+    const company = await this.findById(id);
+    if (!company) throw new NotFoundException('Company not found!');
     return await this.prisma.company.update({ where: { id }, data });
   }
 
   async remove(id: string) {
+    const company = await this.findById(id);
+    if (!company) throw new NotFoundException('Company not found!');
     return await this.prisma.company.delete({ where: { id } });
   }
 }
