+++ backend/src/municipality/municipality.service.ts
@@ -1,4 +1,8 @@
-import { ConflictException, Injectable } from '@nestjs/common';
+import {
+  ConflictException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { CreateMunicipalityDto } from './dto/create-municipality.dto';
 import { UpdateMunicipalityDto } from './dto/update-municipality.dto';
 import { PrismaService } from 'src/prisma/prisma.service';
@@ -8,11 +12,9 @@ export class MunicipalityService {
   constructor(private prisma: PrismaService) {}
 
   async create(data: CreateMunicipalityDto) {
-    const city = await this.prisma.municipality.findUnique({
-      where: { name: data.name },
-    });
+    const municipality = await this.findByName(data.name);
 
-    if (city)
+    if (municipality)
       throw new ConflictException(
         `La commune ${data.name} que vous essayez de créer existe déjà`,
       );
@@ -20,7 +22,7 @@ export class MunicipalityService {
     return await this.prisma.municipality.create({ data });
   }
 
-  async findAll() {
+  async findAll( ) {
     // const { skip, take, cursor, where, orderBy } = params;
     return this.prisma.municipality.findMany({
       // skip,
@@ -32,10 +34,25 @@ export class MunicipalityService {
   }
 
   async findOne(id: string) {
-    return await this.prisma.municipality.findUnique({ where: { id } });
+    const municipality = await this.prisma.municipality.findUnique({
+      where: { id },
+    });
+    if (!municipality) throw new NotFoundException('Municipality not found');
+    return municipality;
+  }
+
+  async findByName(name: string) {
+    const municipality = await this.prisma.municipality.findUnique({
+      where: { name },
+    });
+    if (!municipality) throw new NotFoundException('Municipality not found');
+    return municipality;
   }
 
   async update(id: string, data: UpdateMunicipalityDto) {
+    const municipality = await this.findOne(id);
+    if (!municipality)
+      throw new NotFoundException("This municipality doesn't exist!");
     return await this.prisma.municipality.update({
       where: { id },
       data: { ...data },
@@ -43,6 +60,9 @@ export class MunicipalityService {
   }
 
   async remove(id: string) {
+    const municipality = await this.findOne(id);
+    if (!municipality)
+      throw new NotFoundException("This district doesn't exist!");
     return await this.prisma.municipality.delete({ where: { id } });
   }
 }
